+++ src/components/Categories.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const  Categories = () => (
+    <div>
+      <h2>Under Construction</h2>
+    </div>
+  )
+
+export default Categories;
