--- src/redux/Books/Books.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Books = () => (
-    <div>
-      <h1>All Books</h1>
-    </div>
-  )
-
-export default Books;
