--- src/redux/Categories/Categories.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const  Categories = () => (
-    <div>
-      <h2>Under Construction</h2>
-    </div>
-  )
-
-export default Categories;
